// backend/src/generated/client/runtime/library.d.ts
declare type Options = {
  clientVersion: string
}
// ...
export declare class PrismaClientValidationError extends Error {
  name: string
  clientVersion: string
  constructor(message: string, { clientVersion }: Options)
  get [Symbol.toStringTag](): string
}